Refactor search term processing and TS query construction
Replaced inline TS query construction with a dedicated method in `FilterPreProcessor`. Adjusted `ConceptFilterQueryGenerator` to use `to_tsquery` directly. Modified test request to include a more complex search term.
