+++ dictonaryReqeust.http
@@ -12,4 +12,5 @@ Content-Type: application/json
 POST http://localhost:80/search
 Content-Type: application/json
 
-{"@type":"GeneralQueryRequest","resourceCredentials":{},"query":{"searchTerm":"breast","includedTags":[],"excludedTags":[],"returnTags":"true","offset":0,"limit":10000000},"resourceUUID":null}
+{"@type":"GeneralQueryRequest","resourceCredentials":{},"query":{"searchTerm":"throat hurts acute 8", "includedTags":[],
+  "excludedTags":[],"returnTags":"true","offset":0,"limit":10000000},"resourceUUID":null}
